committing merge

+++ public/js/app.js
@@ -42835,17 +42835,20 @@ if (document.location.href.split('/')[3] == 'new-ltc' || document.location.href.
         township: value.township
       });
       magArray[count].setMap(map);
-      console.log(quickEvict.userId);
-      google.maps.event.addListener(magArray[count], 'mouseover', function (e) {
-        var magistrateId = $(this)[0].areaName.split('magistrate_');
-        injectTooltip(e, magistrateId[1] + '<br>' + $(this)[0].county + '<br>' + $(this)[0].township);
-      });
-      google.maps.event.addListener(magArray[count], 'mousemove', function (e) {
-        moveTooltip(e);
-      });
-      google.maps.event.addListener(magArray[count], 'mouseout', function (e) {
-        deleteTooltip(e);
-      });
+
+      if (quickEvict.userId == 'Administrator') {
+        google.maps.event.addListener(magArray[count], 'mouseover', function (e) {
+          var magistrateId = $(this)[0].areaName.split('magistrate_');
+          injectTooltip(e, magistrateId[1] + '<br>' + $(this)[0].county + '<br>' + $(this)[0].township);
+        });
+        google.maps.event.addListener(magArray[count], 'mousemove', function (e) {
+          moveTooltip(e);
+        });
+        google.maps.event.addListener(magArray[count], 'mouseout', function (e) {
+          deleteTooltip(e);
+        });
+      }
+
       count++;
     });
     autocomplete.addListener('place_changed', function () {
